% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prop_plastic.R
\name{calc_plastic_type_props}
\alias{calc_plastic_type_props}
\title{Create a table of plastic type proportions by year}
\usage{
calc_plastic_type_props(data)
}
\arguments{
\item{data}{A data frame containing plastic waste data.}
}
\value{
A gt table showing plastic type proportions by year.
}
\description{
Calculates the proportion of each plastic type out of the total recorded
plastic waste for each year and returns a formatted gt table.
}
