% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_plastic_by_company.R
\name{plastic_by_company}
\alias{plastic_by_company}
\title{Summarize plastic waste by company}
\usage{
plastic_by_company(data, top_n = 10)
}
\arguments{
\item{data}{A data frame containing plastic waste data.}

\item{top_n}{Number of companies to return.}
}
\value{
A tibble containing parent companies and total plastic waste.
}
\description{
Calculates the total recorded plastic waste for each parent company
and returns the top companies by total plastic waste.
}
