% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plastic_model.R
\name{plastic_model_table}
\alias{plastic_model_table}
\title{Create a gt table for the plastic model}
\usage{
plastic_model_table(model)
}
\arguments{
\item{model}{A fitted lmer model.}
}
\value{
A gt table.
}
\description{
Create a gt table for the plastic model
}
